+++ client/ui.h
@@ -39,7 +39,6 @@ void ShowGraphWindow(void);
 void RepaintGraphWindow(void);
 void PrintAndLogOptions(const char *str[][2], size_t size, size_t space);
 void PrintAndLogEx(logLevel_t level, const char *fmt, ...);
-void SetLogFilename(char *fn);
 void SetFlushAfterWrite(bool value);
 void memcpy_filter_ansi(void *dest, const void *src, size_t n, bool filter);
 
@@ -48,6 +47,8 @@ extern int PlotGridX, PlotGridY, PlotGridXdefault, PlotGridYdefault, GridOffset;
 extern uint32_t CursorCPos, CursorDPos;
 extern bool GridLocked;
 extern bool showDemod;
+
+int searchHomeFilePath(char **foundpath, const char *filename, bool create_home);
 
 extern pthread_mutex_t print_lock;